archiveorigin_backend_api/app/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
orjson==3.10.7
python-ulid==2.7.0
# Optional for signature verification (enabled with VERIFY_SIGNATURES=true)
pynacl==1.5.0
cachetools==5.3.3
ntplib==0.4.0
httpx==0.27.2
pytest==8.3.2
